integration-tests/requirements.txt
pytest==7.4.3
pytest-asyncio==0.21.1
kafka-python-ng
influxdb-client==1.38.0
requests==2.31.0
